move image handle to FormToSaveRecipe component to be able to set empty filename after successul recipe creation
