+++ src/components/ImageInputField.jsx
@@ -1,21 +1,12 @@
-import React, { useState } from "react";
+import React from "react";
 import { styled } from "@mui/material/styles";
 import { Button, Typography } from "@mui/material";
-import utilities from "../modules/utilities";
 
-const ImageInputField = ({ setImage }) => {
-  const [fileName, setFileName] = useState("");
+const ImageInputField = ({ onImageChange, fileName }) => {
   const Input = styled("input")({
     display: "none"
   });
 
-  const handleImage = async (event) => {
-    event.preventDefault();
-    const file = event.target.files[0];
-    file.name && setFileName(file.name);
-    const encodedFile = await utilities.imageEncoder(file);
-    setImage(encodedFile);
-  };
   return (
     <div>
       <label htmlFor="contained-button-file">
@@ -23,7 +14,7 @@ const ImageInputField = ({ setImage }) => {
           id="contained-button-file"
           data-cy="attach-image"
           accept="image/*"
-          onChange={handleImage}
+          onChange={onImageChange}
           name="image"
           multiple
           type="file"
@@ -32,7 +23,12 @@ const ImageInputField = ({ setImage }) => {
           Add Image
         </Button>
       </label>
-      <Typography variant="caption" gutterBottom style={{ marginLeft: "10px" }}>
+      <Typography
+        variant="caption"
+        data-cy="filename"
+        gutterBottom
+        style={{ marginLeft: "10px" }}
+      >
         {fileName}
       </Typography>
     </div>
